Create custom hook to fetch movies
