+++ src/App.js
@@ -12,17 +12,11 @@ import Loader from './components/Main/Loader';
 import ErrorMessage from './components/Main/ErrorMessage';
 import MovieDetails from './components/Main/MovieDetails';
 import Instruction from './components/Main/Instruction';
-
-const KEY = 'c9ea9aa0';
+import useMovies from './custom hooks/useMovies';
 
 export default function App() {
-  const [movies, setMovies] = useState([]);
-  // const [watched, setWatched] = useState([]);
-  const [isLoading, setIsLoading] = useState(false);
-  const [error, setError] = useState('');
   const [query, setQuery] = useState('');
   const [selectedId, setSelectedId] = useState(null);
-
   const [watched, setWatched] = useState(() => {
     const storedValue = localStorage.getItem('watched');
     return JSON.parse(storedValue);
@@ -48,49 +42,7 @@ export default function App() {
     localStorage.setItem('watched', JSON.stringify(watched));
   }, [watched]);
 
-  useEffect(() => {
-    const controller = new AbortController();
-
-    const fetchMovies = async () => {
-      try {
-        setIsLoading(true);
-        setError('');
-        const res = await fetch(
-          `http://www.omdbapi.com/?apikey=${KEY}&s=${query}`,
-          { signal: controller.signal },
-        );
-
-        if (!res.ok) throw new Error('Something went wrong with fetching movies.');
-
-        const data = await res.json();
-
-        if (data.Response === 'False') throw new Error('Movie not found');
-
-        setMovies(data.Search);
-        setError('');
-        setIsLoading(false);
-      } catch (err) {
-        if (err.name === 'AbortError') {
-          setIsLoading(true);
-        } else {
-          setError(err.message);
-          setIsLoading(false);
-        }
-      }
-    };
-
-    if (!query.length) {
-      setMovies([]);
-      setError('');
-    }
-
-    handleCloseMovie();
-    fetchMovies();
-
-    return () => {
-      controller.abort();
-    };
-  }, [query]);
+  const { movies, isLoading, error } = useMovies(query, handleCloseMovie);
 
   return (
     <>
